+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Playback can be stopped after the current track finishes with the "Stop after current" menu item.
+
 ### Changed
 - Clicking the menubar buttons now sends `NSWorkspaceDidActivateApplication` notifications to ensure [MediaKeyTap](https://github.com/nhurden/MediaKeyTap) switches kmbmpdc to most recently active application.
 - ID3v2 cover art parsing changed to use [imeji](https://github.com/arttuperala/imeji).